refactor: refactor all record types validation
