+++ crates/bindizr-service/src/record/validation/record_value/soa.rs
@@ -1,0 +1,13 @@
+pub(super) struct SoaRecordValue<'a> {
+    value: &'a str,
+}
+
+impl<'a> SoaRecordValue<'a> {
+    pub(super) fn parse(value: &'a str) -> Self {
+        Self { value }
+    }
+
+    pub(super) fn canonical(&self) -> String {
+        self.value.to_string()
+    }
+}
